[CI] Web  upload factorizing
Update esp web tools
Simplify scripts
