+++ .github/workflows/build.yml
@@ -77,7 +77,7 @@ jobs:
           - "esp32s3-dev-c1-ble"
           - "esp32c3-dev-m1-ble"
           - "airm2m_core_esp32c3"
-          - "lolin_c3_mini"
+          - "esp32c3_lolin_mini"
           - "esp32c3-m5stamp"
           - "thingpulse-espgateway"
           - "theengs-bridge"
