+++ docs/.vuepress/styles/index.styl
@@ -1,0 +1,11 @@
+// index.styl
+div.table-generated-page
+  main
+    div.content__default,
+    footer.page-edit,
+    div.page-nav
+      // Display full with the API pages
+      max-width: 65%
+
+.sidebar
+  z-index 6
