+++ environments.ini
@@ -1716,7 +1716,7 @@ build_flags =
 custom_description = BLE gateway on Espressiv ESP32-C3-DevKitC-02 without logging
 
 ; Wemos Lolin C3 mini v2.1.0
-[env:lolin_c3_mini]
+[env:esp32c3_lolin_mini]
 platform = ${com.esp32_platform}
 board = lolin_c3_mini
 board_build.partitions = min_spiffs.csv
@@ -1746,10 +1746,10 @@ build_flags =
 custom_description = BLE gateway on ESP32C3
 custom_hardware = Wemos Lolin C3 mini v2.1.0
 
-[env:lolin_c3_mini_with_serial]
-extends = env:lolin_c3_mini
+[env:esp32c3_lolin_mini_with_serial]
+extends = env:esp32c3_lolin_mini
 build_flags =
-  ${env:lolin_c3_mini.build_flags}
+  ${env:esp32c3_lolin_mini.build_flags}
   '-DLOG_LEVEL=LOG_LEVEL_NOTICE' ;default
   '-DWM_DEBUG_LEVEL=1' ;DEBUG_NOTIFY=1, default value
 ; Serial output via USB
